// web/src/api/generated/PayoutDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Amount } from "./Amount";
import type { NodeId } from "./NodeId";
import type { Tier } from "./Tier";

/**
 * One node payout inside a block.
 */
export type PayoutDto = { tier: Tier, node: NodeId | null, address: string, amount: Amount, };